requirements.txt
Flask==3.0.2
langchain==0.1.7
langchain_community==0.0.20
langchain_core==0.1.23
langchain_openai==0.0.6
python-dotenv==1.0.1
Requests==2.31.0
serpapi==0.1.5
together==0.2.11
trafilatura==1.7.0